make it a little easier to debug by having the second param use a different type annotation

+++ test_fixtures/simple_typing.py
@@ -1,7 +1,7 @@
 
 
-def add(a: int, b: int) -> int:
-    c = a + b # type: int
+def add(a: int, b: float) -> int:
+    c = a + b  # type: int
 
     return c
